requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-jose[cryptography]
bcrypt
python-multipart
jinja2
httpx
python-dotenv
